Moved IEditorProvider to the AvalonStudio.Shell namespace.
