+++ AvalonStudio/AvalonStudio.Extensibility/Shell/ExportEditorProviderAttribute.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Composition;
+
+namespace AvalonStudio.Shell
+{
+    [AttributeUsage(AttributeTargets.Class)]
+    public class ExportEditorProviderAttribute : ExportAttribute
+    {
+        public ExportEditorProviderAttribute()
+            : base(typeof(IEditorProvider))
+        {
+        }
+    }
+}
