+++ AvalonStudio/AvalonStudio.Extensibility/Shell/IEditorProvider.cs
@@ -1,12 +1,11 @@
 ﻿using AvalonStudio.Documents;
 using AvalonStudio.Projects;
 
-namespace AvalonStudio.Extensibility.Editor
+namespace AvalonStudio.Shell
 {
     public interface IEditorProvider
     {
         bool CanEdit(ISourceFile file);
-
         IFileDocumentTabViewModel CreateViewModel(ISourceFile file);
     }
 }
